Fix errors with boolean types in method return and flow types on iOS
Fixes "_TypeError (type 'int' is not a subtype of type 'bool?' in type cast)".
We now serialize boolean values because they are ints when received via Kotlin -> Dart leading to these casting errors.
For simplicity the serialization is added for both directions and for Android and iOS, although only Kotlin -> Dart and iOS would be necessary.